Add license

+++ app/client/managers/models/IntervalManager.java
@@ -1,3 +1,24 @@
+/*
+ * WBI Indicator Explorer
+ *
+ * Copyright 2015 Sebastian Nogara <snogaraleal@gmail.com>
+ *
+ * This file is part of WBI.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package client.managers.models;
 
 import java.util.ArrayList;
@@ -58,8 +79,8 @@ public class IntervalManager implements Manager {
         @Override
         public int hashCode() {
             int hash = 3;
-            hash = 37 * hash + (int)(startYear ^ (startYear >>> 32));
-            hash = 37 * hash + (int)(endYear ^ (endYear >>> 32));
+            hash = 37 * hash + (startYear ^ (startYear >>> 32));
+            hash = 37 * hash + (endYear ^ (endYear >>> 32));
             return hash;
         }
     }
